fix: 修复浏览器关闭和类型构建问题
修复IBrowser中chrome.close的可选调用问题
将doc2tsConfigContent转换为Uint8Array类型写入文件
更新enum构建时的key引号处理
优化log.clear方法的TTY检查
添加iconv-lite依赖并移除无用配置路径
确保在Doc2Ts处理完成后关闭浏览器实例

+++ src/scripts/manage.ts
@@ -174,7 +174,7 @@ export default class Manage {
     } else {
       await this.initBranchname()
     }
-    fs.writeFileSync(getRootFilePath(CONFIG_PATH), this.doc2tsConfigContent)
+    fs.writeFileSync(getRootFilePath(CONFIG_PATH), this.doc2tsConfigContent as Uint8Array)
   }
 
   async getBranch() {
